dloper-os-pro/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pyyaml==6.0.1
PyJWT==2.8.0
python-multipart==0.0.9
psutil==5.9.8
docker==6.1.3
aiofiles==23.2.1
